Remove the views and router, and remove vue-router dependency
