+++ package.json
@@ -30,8 +30,7 @@
     "@lumino/default-theme": "^0.4.2",
     "@lumino/widgets": "^1.13.2",
     "core-js": "^3.6.5",
-    "vue": "^2.6.11",
-    "vue-router": "^3.3.4"
+    "vue": "^2.6.11"
   },
   "devDependencies": {
     "@cypress/code-coverage": "^3.8.1",
